Move Storage to inner include
This should make Storage not visible from duckdb.hpp, so that less conflicts like #6504
would happen. Broader idea would be that functions tagged DUCKDB_API should be visible
from duckdb.hpp, the rest are just implementation details that can potentially hidden.
