+++ src/include/duckdb/main/connection.hpp
@@ -8,8 +8,6 @@
 
 #pragma once
 
-#include "duckdb/common/enums/profiler_format.hpp"
-#include "duckdb/common/serializer/buffered_file_writer.hpp"
 #include "duckdb/common/winapi.hpp"
 #include "duckdb/function/udf_function.hpp"
 #include "duckdb/main/materialized_query_result.hpp"
